don't wait for child if there are resources in the wild
Make the logic clear, regarding what kind of resources are involved
and who shall handle their cleanup (we or caller). That is:

- we may allocate two kind of system resources: I/O (pipes), and
  process (child).
- block is to guard I/O resources (close at the end)
- we regard command arguments (ie. child is doing exec) as wish
  to guard process resources (wait for child)
- we have a constraint: I/O cleanup must precede process cleanup (ie.
  no point in giving the caller pipes to a dead child)
- therefore:
  1. if block is given, we unconditionally imply closing
     pipes at the end
  2. if arguments given, we wait for child if we can, ie. no open pipes exist
- if 1.: we return final result, ie. process status
- if 2.: we return handles to outstanding resources (array of open pipes +
  child pid)

In particular, if there are no cmd args, we do not wait -- not even
if I/O is guarded. The block is not considered authorative wrt. child
handling :)

Simple, eh?