Have run! always return [..fd.., pid], and don't wait for child
This makes #run! a more convenient as a programming interface,
and let #run be there for direct use with all of its convenience
shortcuts.

There is no point in maintaining two custom event handling logics,
a permissive and a restictive one. Instead You get a raw variant
with not much cleverness and one with all cleverness I could find out.

Note also that since we have the exec-failure-handling stuff in place,

  pid, = run or exec "cmd"

cannot be regarded as the variant which execs but doesn't waits,
as this misses out exec protection. Now we can do

  run! "cmd"

to get at non-waiting exec (with fault handling).