move the whole stuff into the internal class
We really need to become fully stateful...

Wth a simple run-program-read-output scenario,
when I/O blowed up within the block, child wasn't waited
in the exception handling code, so it remained there as a zombie
forever.

However, we have to be cautious about waiting for child as a cleanup
action: it can get us hung, with eg. a daemonic child. What seems
to be sane is the following super-simple principle: wait in exception
handler if we waited for child normally as well. The catch is, "if we waited
for child normally" was hard to track down, as the wait/no-wait logic was
splintered over along the event chain, rendered into a simple Boolean value
only by trouble free termination. It was like so, despite the fact that a
static analysis of the arguments determines whether to wait or not!

So now we parse arguments in #initialize, set up hi-level logic /
protection in interface methods, then call core function for which
all flow control turnpoints are readily available as instance data.