Revert "monkey-patch unwaited child pid onto exceptions"
This reverts commit 7199b512f705b65e9d683228ca1c6beea9ee3e21.

Instead of just popping off HEAD, I record hereby a reverse commit
to take the opportunity to tell why it was not a good idea.

An exception originated from us is either catched by the generic
exception handler code n levels higher, or by a dedicated handler
which is there just to guard this specific invocation.

First case does not profit from odd exception methods at all, as
it's _generic_.

Second case means that exceptions are made part of the API. User have
to be aware of what exceptions may come from us, and write code matching
that. That's no good design.

OK it was a nice try pointing out that resource can leak by some scenarios...
but we have to go and tackle this issue in a different way, which we'll do
now.