make redirections proper objects
This was enforced by a bug in #want_wait?:
it relied on the initial format of @redir, which was
overloaded later on with runtime information, changing
the result. So eg. the following failed due to bogusly
getting into wait mode:

  f,pid = run("whoami", 1); puts f.read'

With Redir instances, it's possible to properly separate
static and dynamic attributes of redirections.